[FIX] chat invite fixes
