+++ srcs/frontend/transcendence-app/src/chat/chatUI.ts
@@ -328,29 +328,57 @@ export class ChatUIManager {
     document.removeEventListener('click', this.handleOutsideClick);
   }
 
+  private setupUsernameClickHandlers(): void {
+    const usernameElements = document.querySelectorAll('b[data-username]:not([data-handler-added])');
+    
+    usernameElements.forEach(element => {
+      const username = element.getAttribute('data-username');
+      if (username) {
+        element.addEventListener('click', async (event) => {
+          const { userModal } = await import('./userModal');
+          userModal.showUserModal(username, event as MouseEvent);
+        });
+        element.setAttribute('data-handler-added', 'true');
+      }
+    });
+  }
+
   // ============================================================================ //
   // GAME INVITE UI                                                               //
   // ============================================================================ //
 
-  displayGameInvite(senderId: string, gameId: string, userId: string, notifId: string): void {
+  displayGameInvite(senderId: string, gameId: string, userId: string, notifId: string, isMe: boolean = false, senderUsername?: string): void {
     try {
       console.log('Displaying game invite in chat UI from', senderId, 'for game', gameId);
 
       const messagesDiv = document.getElementById('sidebar-chat-messages');
       if (!messagesDiv) {
-        console.warn('Messages div not found, cannot display game invite in chat');
         return;
       }
 
       const inviteId = `game-invite-${gameId}-${Date.now()}`;
+      const alignClass = isMe ? 'justify-end' : 'justify-start';
+      const bgColor = isMe ? 'bg-blue-100 text-[#1a2740]' : 'bg-yellow-100 text-[#1a2740]';
+      const borderColor = isMe ? 'border-blue-300' : 'border-yellow-300';
+      const inviteText = isMe ? 'You sent a game invite' : 'You received a game invite';
+      
+      const displayName = isMe ? 'Me' : (senderUsername || 'Unknown');
+      const usernameDisplay = isMe || displayName === 'Me' 
+        ? `<b>${displayName}:</b>`
+        : `<b class="cursor-pointer hover:text-blue-400 hover:underline transition-colors duration-200" data-username="${displayName}">${displayName}:</b>`;
 
       messagesDiv.innerHTML += `
-        <div class="mb-2 flex justify-start">
-          <div class="bg-yellow-100 text-[#1a2740] rounded-lg px-3 py-2 max-w-[75%] shadow border border-yellow-300 text-sm">
-            <p class="mb-3">You received a game invite</p>
-            <div class="flex gap-2">
-              <button id="acceptGameBtn-${inviteId}" class="bg-green-600 hover:bg-green-700 text-white text-xs px-3 py-1 rounded">Accept</button>
-              <button id="declineGameBtn-${inviteId}" class="bg-red-600 hover:bg-red-700 text-white text-xs px-3 py-1 rounded">Decline</button>
+        <div class="mb-2 flex ${alignClass}">
+          <div class="max-w-[75%]">
+            <div class="mb-1 ${isMe ? 'text-right' : 'text-left'}">
+              <span class="text-xs text-gray-400">${usernameDisplay}</span>
+            </div>
+            <div class="${bgColor} rounded-lg px-3 py-2 shadow border ${borderColor} text-sm">
+              <p class="mb-3">${inviteText}</p>
+              <div class="flex gap-2">
+                <button id="acceptGameBtn-${inviteId}" class="bg-green-600 hover:bg-green-700 text-white text-xs px-3 py-1 rounded">Accept</button>
+                <button id="declineGameBtn-${inviteId}" class="bg-red-600 hover:bg-red-700 text-white text-xs px-3 py-1 rounded">Decline</button>
+              </div>
             </div>
           </div>
         </div>
@@ -360,10 +388,81 @@ export class ChatUIManager {
 
       setTimeout(() => {
         this.setupGameInviteButtons(inviteId, gameId, userId, notifId);
+        this.setupUsernameClickHandlers();
       }, 0);
 
     } catch (error) {
       console.error('Error displaying game invite in chat UI:', error);
+    }
+  }
+
+  displayGameInviteResponded(isMe: boolean = false, senderUsername?: string): void {
+    try {
+      const messagesDiv = document.getElementById('sidebar-chat-messages');
+      if (!messagesDiv) {
+        console.warn('Messages div not found, cannot display responded game invite');
+        return;
+      }
+
+      const alignClass = isMe ? 'justify-end' : 'justify-start';
+      const inviteText = isMe ? 'You sent a game invite' : 'Invite To Play';
+      const statusText = isMe ? 'Invitation sent.' : 'Invitation responded.';
+
+      const displayName = isMe ? 'Me' : (senderUsername || 'Unknown');
+      const usernameDisplay = isMe || displayName === 'Me' 
+        ? `<b>${displayName}:</b>`
+        : `<b class="cursor-pointer hover:text-blue-400 hover:underline transition-colors duration-200" data-username="${displayName}">${displayName}:</b>`;
+
+      messagesDiv.innerHTML += `
+        <div class="mb-2 flex ${alignClass}">
+          <div class="max-w-[75%]">
+            <div class="mb-1 ${isMe ? 'text-right' : 'text-left'}">
+              <span class="text-xs text-gray-400">${usernameDisplay}</span>
+            </div>
+            <div class="bg-gray-200 text-gray-600 rounded-lg px-3 py-2 shadow border border-gray-300 text-sm opacity-70 cursor-not-allowed">
+              <p class="mb-2 font-medium">${inviteText}</p>
+              <p class="mb-2 italic">${statusText}</p>
+            </div>
+          </div>
+        </div>
+      `;
+
+      messagesDiv.scrollTop = messagesDiv.scrollHeight;
+
+      setTimeout(() => {
+        this.setupUsernameClickHandlers();
+      }, 0);
+
+    } catch (error) {
+      console.error('Error displaying responded game invite:', error);
+    }
+  }
+
+  async checkNotificationStatus(userId: string, notifId: string): Promise<boolean> {
+    try {
+      const token = chatState.getAuthToken();
+      const response = await fetch(`https://localhost:4242/notifications/${userId}`, {
+        headers: {
+          'Authorization': `Bearer ${token}`
+        },
+        credentials: 'include'
+      });
+
+      if (!response.ok) {
+        console.warn('Could not fetch notification status');
+        return false;
+      }
+
+      const data = await response.json();
+      
+      const notification = data.notifications?.find((notif: any) => {
+        return notif.id.toString() === notifId.toString();
+      });
+      
+      return notification ? notification.is_read === "read" : false;
+    } catch (error) {
+      console.error('Error checking notification status:', error);
+      return false;
     }
   }
