+++ srcs/frontend/transcendence-app/src/chat/chatWebSocket.ts
@@ -186,6 +186,26 @@ export class ChatWebSocketManager {
       const data = JSON.parse(event.data);
       console.log('Received WebSocket message:', data);
 
+      if (data.message) {
+        try {
+          const messageData = JSON.parse(data.message);
+          if (messageData.type === 'game.invite') {
+            console.log('Game invite found in message:', messageData);
+            chatUI.displayGameInvite(
+              messageData.senderId || '',
+              messageData.gameId || '',
+              messageData.receiverId || '',
+              messageData.notifId || '',
+              false,
+              messageData.username || 'Unknown'
+            );
+            return;
+          }
+        } catch {
+
+        }
+      }
+
       const messageHandlers: Record<string, Function> = {
         'user_blocked': () => userBlocking.handleUserBlocked(data.blocked_by_user_id, data.blocked_by_username),
         'user_blocked_by_me': () => userBlocking.handleUserBlockedByMe(data.blocked_user_id),
@@ -195,7 +215,6 @@ export class ChatWebSocketManager {
           console.log('Friend request accepted, rejoining all available rooms');
           await this.joinAllAvailableRooms();
         },
-        'game.invite': () => chatUI.displayGameInvite(data.senderId, data.gameId, data.receiverId, data.notifId),
       };
       
       const handler = messageHandlers[data.type];
@@ -223,7 +242,7 @@ export class ChatWebSocketManager {
         chatMessages.showErrorMessage(messageText);
         console.warn('Server message:', messageText);
       } else {
-        console.error('Error parsing WebSocket message:', error, event.data);
+        
       }
     }
   }
@@ -243,7 +262,7 @@ export class ChatWebSocketManager {
     return chatState.currentWs !== null && chatState.currentWs.readyState === WebSocket.OPEN;
   }
 
-  sendMessage(scope: 'group' | 'dm', room: number, message: string): boolean {
+  sendMessage(scope: 'group' | 'dm', room: number, message: any): boolean {
     if (this.isWebSocketConnected() && chatState.currentWs) {
       chatState.currentWs.send(JSON.stringify({
         action: 'send',
@@ -251,6 +270,16 @@ export class ChatWebSocketManager {
         room,
         message
       }));
+      
+      try {
+        if (JSON.stringify(message).includes('game.invite')) {
+          chatUI.displayGameInviteResponded(true, 'Me');
+          return true;
+        }
+      } catch (error) {
+
+      }
+      
       return true;
     }
     console.warn('Cannot send message: WebSocket not connected');
